+++ ResumableFunctions.Handler/InOuts/IResumableFunctionsSettings.cs
@@ -9,7 +9,7 @@ namespace ResumableFunctions.Handler.InOuts
         public IGlobalConfiguration HangfireConfig { get; }
         public DbContextOptionsBuilder WaitsDbConfig { get; }
         public string CurrentServiceUrl { get; }
-        public int CurrentServiceId { get; internal set; }
+        public long CurrentServiceId { get; internal set; }
 
         public IDistributedLockProvider DistributedLockProvider { get; }
         public string[] DllsToScan { get; }
