+++ ResumableFunctions.Handler/Testing/TestSettings.cs
@@ -29,7 +29,7 @@ namespace ResumableFunctions.Handler.Testing
 
         public bool ForceRescan { get; set; } = true;
         public string CurrentDbName { get; set; }
-        public int CurrentServiceId { get; set; } = -1;
+        public long CurrentServiceId { get; set; } = -1;
 
         public IDistributedLockProvider DistributedLockProvider =>
             new WaitHandleDistributedSynchronizationProvider();
