+++ ResumableFunctions.Handler/UiService/InOuts/MethodGroupInfo.cs
@@ -5,7 +5,7 @@ namespace ResumableFunctions.Handler.UiService.InOuts
     public record MethodInGroupInfo(string ServiceName, WaitMethodIdentifier Method,string GroupUrn);
 
     public record MethodGroupInfo(
-        int Id, string URN, int MethodsCount, int ActiveWaits, int CompletedWaits, int CanceledWaits, DateTime Created)
+        long Id, string URN, int MethodsCount, int ActiveWaits, int CompletedWaits, int CanceledWaits, DateTime Created)
     {
         public int AllWaitsCount => ActiveWaits + CompletedWaits + CanceledWaits;
     }
